chore: rename to strictMarksPlugin
